# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Find My MUA API",
        "description": "Dokumentasi lengkap API untuk aplikasi Find My MUA",
        "version": "1.0"
    },
    "paths": {
...
        "/api/customer/bookings": {
            "get": {
                "tags": [
                    "Booking"
                ],
                "summary": "Lihat semua booking milik customer",
                "operationId": "a3442a0a9acd411936ad4f5673735660",
                "responses": {
                    "200": {
                        "description": "Daftar booking"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Booking"
                ],
                "summary": "Buat booking baru",
                "operationId": "753e234378345871167a3a04b155d307",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "mua_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "service_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-07-01"
                                    },
                                    "time": {
                                        "type": "string",
                                        "example": "14:00"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "example": "transfer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking berhasil dibuat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/mua/bookings": {
            "get": {
                "tags": [
                    "Booking"
                ],
                "summary": "Lihat semua booking untuk MUA",
                "operationId": "3af2ad63f8140f7ab534cafaf1ff17a5",
                "responses": {
                    "200": {
                        "description": "List booking untuk MUA"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/mua/{id}": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Lihat profil publik MUA",
                "operationId": "b8a4625af8a6d80324d6675f5b9e9abc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profil publik MUA"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}